requirements.txt
requests
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
